Implement and add option for fast kmer processing.
The new implementation is very fast and allow us to handle mismatch
kernels on a new scale e.g. k=12, m=2 in less than 36s on my laptop.

This commit is not well tested and is intended for quick deployment
with feedback loop.
TODO: test more kmer_processor and reformat benchmark which looks
messy.